+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [0.21.3] - 2025-07-27
 ### Added
 - Adds `rounded_corners` option, for round corners of the bounding box.
 ### Fixed